+++ RELEASES.md
@@ -1,4 +1,5 @@
 * Add `Rational::cmp_abs` method.
+* Add `Float::sum` and `Complex::sum` methods.
 * Add `from_raw`, `into_raw`, `as_raw` and `as_raw_mut` to
   `RandState`.